requirements.txt
streamlit
sentence-transformers
faiss-cpu
pandas
numpy
openai
transformers
torch
tqdm
plotly
networkx
